+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.11.0] - 2021-08-08
+### Added
+- Live reporting of progress parsing files (below the component)
+
+### Changed
+- Reduced minimum required Rhino 6 version to `6.10` or greater
+
 ## [0.10.0] - 2021-07-21
 ### Added
 - A bake-friendly layer name to the `Report` output parameter